+++ jest.config.ts
@@ -1,11 +1,7 @@
-const e2e = process.env.TEST_E2E,
-  unit = process.env.TEST_UNIT;
-
 export default {
-  preset: unit ? undefined : "jest-puppeteer",
-  testMatch: [`**/tests/${e2e ? "e2e" : unit ? "!(e2e)" : "*"}.test.ts`],
+  preset: "jest-puppeteer",
   transform: { "^.+\\.ts?$": "ts-jest" },
-  collectCoverage: !Boolean(e2e),
+  collectCoverage: true,
   coverageDirectory: "<rootDir>/tests/coverage",
   collectCoverageFrom: ["src/*.ts"],
 };
